scripts/requirements.txt
pdf2image>=1.17.0
Pillow>=12.0.0
openai>=2.26.0
requests>=2.31.0
pydantic>=2.12.0
numpy>=1.24.0
mcp[cli]>=1.0.0
